# Conflicts:
#	src/main/java/com/br/luggycar/api/services/VehicleService.java

+++ src/main/java/com/br/luggycar/api/services/VehicleService.java
@@ -1,12 +1,17 @@
 package com.br.luggycar.api.services;
 
+import com.br.luggycar.api.dtos.response.ClientResponse;
 import com.br.luggycar.api.dtos.response.VehicleResponse;
 import com.br.luggycar.api.entities.Category;
 import com.br.luggycar.api.entities.Client;
 import com.br.luggycar.api.entities.Vehicle;
+import com.br.luggycar.api.enums.rent.RentStatus;
+import com.br.luggycar.api.exceptions.ResourceDatabaseException;
 import com.br.luggycar.api.exceptions.ResourceExistsException;
 import com.br.luggycar.api.exceptions.ResourceNotFoundException;
 import com.br.luggycar.api.repositories.CategoryRepository;
+import com.br.luggycar.api.repositories.ClientRepository;
+import com.br.luggycar.api.repositories.RentRepository;
 import com.br.luggycar.api.repositories.VehicleRepository;
 import com.br.luggycar.api.dtos.requests.VehicleRequest;
 import org.springframework.beans.BeanUtils;
@@ -25,13 +30,10 @@ public class VehicleService {
     private VehicleRepository vehicleRepository;
     @Autowired
     private CategoryRepository categoryRepository;
+    @Autowired
+    private RentRepository rentRepository;
 
     public VehicleResponse createVehicle(VehicleRequest vehicleRequest) {
-
-        Optional<Vehicle> existingVehicle = vehicleRepository.findByPlate(vehicleRequest.getPlate());
-        if (existingVehicle.isPresent()) {
-            throw new ResourceExistsException("Já existe um veículo cadastrado com essa placa.");
-        }
 
         Category category = categoryRepository.findByName(vehicleRequest.categoryName())
                 .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
@@ -44,7 +46,6 @@ public class VehicleService {
         Vehicle savedVehicle = vehicleRepository.save(vehicle);
 
         return new VehicleResponse(savedVehicle);
-
 
     }
 
@@ -61,6 +62,12 @@ public class VehicleService {
         Vehicle vehicle = vehicleRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Veículo não encontrado"));
 
+        List<RentStatus> activeStatuses = List.of(RentStatus.PENDING, RentStatus.IN_PROGRESS);
+
+        if (rentRepository.existsByVehicleIdAndStatusIn(id,activeStatuses)) {
+            throw new ResourceExistsException("veículo com aluguel em curso não pode ser editado.");
+        }
+
         BeanUtils.copyProperties(vehicleRequest, vehicle, "id", "registrationDate");
 
         Vehicle updatedVehicle = vehicleRepository.save(vehicle);
@@ -69,21 +76,36 @@ public class VehicleService {
     }
 
     public void deleteVehicle(Long id) {
+
+        List<RentStatus> activeStatuses = List.of(RentStatus.PENDING, RentStatus.IN_PROGRESS);
+
+        if (rentRepository.existsByVehicleIdAndStatusIn(id,activeStatuses)) {
+            throw new ResourceExistsException("veículo com aluguel em curso não pode ser excluido.");
+        }
+
         vehicleRepository.deleteById(id);
     }
-
 
     public Optional<VehicleResponse> findVehicleById(Long id) {
         return vehicleRepository.findById(id)
                 .map(VehicleResponse::new);
     }
 
-    public VehicleResponse getByPlate(String plate) {
-        Vehicle vehicle = vehicleRepository.findByPlate(plate)
-                .orElseThrow(() -> new ResourceNotFoundException("Veículo não encontrado com a placa: " + plate));
+    public List <VehicleResponse>getAvailableVehicles() {
 
-        return new VehicleResponse(vehicle);
+        try {
+            List<Vehicle> allVehicles = vehicleRepository.findAll();
+            List<RentStatus> activeStatuses = List.of(RentStatus.PENDING, RentStatus.IN_PROGRESS);
+            List<Vehicle> rentedVehicles = rentRepository.findRentedVehiclesByStatusIn(activeStatuses);
+            allVehicles.removeAll(rentedVehicles);
+
+            return allVehicles.stream()
+                    .map(VehicleResponse::new)
+                    .collect(Collectors.toList());
+
+        } catch (
+                ResourceDatabaseException e) {
+            throw new ResourceDatabaseException("Erro ao buscar os veiculos diponiveis para aluguel no banco de dados", e);
+        }
     }
-
-
 }
